pyproject.toml
[project]
name = "audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.14"
dependencies = [
    "aiofiles==24.1.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "asyncio==4.0.0",
    "brotli==1.2.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "fastapi==0.127.0",
    "ffmpy==1.0.0",
    "filelock==3.20.1",
    "fsspec==2025.12.0",
    "gradio==6.2.0",
    "gradio-client==2.0.2",
    "groovy==0.1.2",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.2.3",
    "idna==3.11",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "nicegui>=3.4.1",
    "numpy==2.4.0",
    "orjson==3.11.5",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.0.0",
    "pyaudio==0.2.14",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings>=2.7.1",
    "pydub==0.25.1",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests>=2.32.5",
    "rich==14.2.0",
    "safehttpx==0.1.7",
    "semantic-version==2.10.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sounddevice==0.5.3",
    "soundfile>=0.13.1",
    "starlette==0.50.0",
    "tomlkit==0.13.3",
    "tqdm==4.67.1",
    "typer==0.20.1",
    "typer-slim==0.20.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "uvicorn==0.40.0",
    "websockets>=15.0.1",
]
